+++ backend/src/controllers/shipmentsControllers.js
@@ -18,8 +18,13 @@ export const getShipmentsById = async (req, res) => {
 }
 
 export const deleteShipmentsById = async (req, res) => {
-    await Shipment.deleteOne({_id: req.params.id})
-    res.status(200).send("Done")
+    let shipment = await Shipment.findById(req.params.id)
+    if(shipment !=null) {
+        await Shipment.deleteOne({_id: req.params.id})
+        return res.status(200).send("Done")
+    }
+
+    res.status(400).send("Not found")
 };
 
 export const addShipment = async (req, res) => {
